fixed  printTrafficPerClientByTimeInterval() methods logic and datarate logic

+++ src/Packet.java
@@ -1,14 +1,14 @@
 
 public class Packet {
-	final int bytes;
+	final long bytes;
 	final long timeStamp;
 
-	public Packet(int bytes, long timeStamp) {
+	public Packet(long bytes, long timeStamp) {
 		this.bytes = bytes;
 		this.timeStamp = timeStamp;
 	}
 
-	public int getBytes() {
+	public long getBytes() {
 		return this.bytes;
 	}
